Xfwm4: improve accessibility (wider bottom edge)
Since Xfwm have never expressed much interest in virtual resize
grips extending away from the window area, and need to support non-
composited configurations anyway, this usability fix is needed in
Numix. Single-pixel borders are not enough.

This commit adds a 5px bottom border and big corner resize grips to
Xfwm allow windows to be resized more easily. This improves matters
significantly for tablet and mouse/touchpad users.

The corner resize grips are from NeXTSTEP, or more accurately its
clones on Linux such as Window Maker. The bottom border as a whole
is styled to match the rest of the theme.

Closes numixproject/numix-gtk-theme#100.