Updates Makefile to replace Ruby Sass with SassC
This commit replaces the use of the Ruby Sass scss command with SassC's sassc command. Due to the more limited functionality of sassc, output directories must be created prior to running the command and only one file may be processed at a time unlike the many-to-many mode functionality of the scss command.